Cleanup copy of repositories from localStorage to IndexedDB
